Significant cleanup of codebase
Fixed broken tests; Added tests for TexRunner and spaces (not working).
Cleaned up documentation across the board.

Things look very exciting at the moment. Unfortunately there may be just
a little too much magic. It's kind of like Django for plotting... but
done right (Django is fine if you do it the Django way, which is not
communicated so well in the manual)

Won't be around the next couple of days, so will use this to note what
needs done:
 * add TODO! (this is a one off...)
 * Rewrite plot
 * write ROADMAP
 * write documentation (something to show is more important!)
 * YODA wrapper! (talk about plugin architecture... very very simple
   idea)
 * write consoleRunner
 * simple box model? (build by refactor)